feat: improve local environment * update docker deployment * add init-app-service * update readme * add .env example
