+++ .env.example
@@ -1,0 +1,4 @@
+ELASTICSEARCH_NODE=http://localhost:9200
+ELASTICSEARCH_USERNAME=elastic-dev
+ELASTICSEARCH_PASSWORD=admin
+ELASTICSEARCH_INDEX=politicians
