+++ src/app.module.ts
@@ -6,6 +6,7 @@ import { HealthModule } from './health/health.module';
 import { SearchModule } from './common/search.module';
 import { BulkModule } from './bulk/bulk.module';
 import { StatisticsModule } from './statistics/statistics.module';
+import { AppService } from './common/services/app.service';
 
 @Module({
   imports: [
@@ -17,5 +18,6 @@ import { StatisticsModule } from './statistics/statistics.module';
     StatisticsModule,
   ],
   controllers: [AppController],
+  providers: [AppService],
 })
 export class AppModule {}
