+++ src/politicians/services/politicians.service.ts
@@ -11,128 +11,7 @@ import { Politician } from '../entities/politician.entity';
 export class PoliticiansService {
   constructor(private readonly elasticsearchService: ElasticsearchService) {}
 
-  public async createIndex() {
-    const checkIndex = await this.elasticsearchService.indices.exists({
-      index: politiciansIndex,
-    });
-
-    if (!checkIndex) {
-      await this.elasticsearchService.indices.create({
-        index: politiciansIndex,
-        body: {
-          mappings: {
-            properties: {
-              fullName: {
-                type: 'text',
-                fields: {
-                  keyword: {
-                    type: 'keyword',
-                    ignore_above: 256,
-                  },
-                },
-              },
-              party: {
-                type: 'text',
-                fields: {
-                  keyword: {
-                    type: 'keyword',
-                    ignore_above: 256,
-                  },
-                },
-              },
-              partyFilter: {
-                type: 'text',
-                fields: {
-                  keyword: {
-                    type: 'keyword',
-                    ignore_above: 256,
-                  },
-                },
-              },
-              gender: {
-                type: 'text',
-                fields: {
-                  keyword: {
-                    type: 'keyword',
-                    ignore_above: 256,
-                  },
-                },
-              },
-              occupation: {
-                type: 'text',
-                fields: {
-                  keyword: {
-                    type: 'keyword',
-                    ignore_above: 256,
-                  },
-                },
-              },
-              occupationFilter: {
-                type: 'text',
-                fields: {
-                  keyword: {
-                    type: 'keyword',
-                    ignore_above: 256,
-                  },
-                },
-              },
-              community: {
-                type: 'text',
-                fields: {
-                  keyword: {
-                    type: 'keyword',
-                    ignore_above: 256,
-                  },
-                },
-              },
-              institution: {
-                type: 'text',
-                fields: {
-                  keyword: {
-                    type: 'keyword',
-                    ignore_above: 256,
-                  },
-                },
-              },
-              baseSalary: {
-                type: 'long',
-              },
-              monthlyPayment: {
-                type: 'long',
-              },
-              yearlyPayment: {
-                type: 'long',
-              },
-              complementSalary: {
-                type: 'long',
-              },
-              extraPay: {
-                type: 'long',
-              },
-              otherAllowancesAndIndemnities: {
-                type: 'long',
-              },
-              threeYearsSalary: {
-                type: 'long',
-              },
-              observations: {
-                type: 'text',
-                fields: {
-                  keyword: {
-                    type: 'keyword',
-                    ignore_above: 256,
-                  },
-                },
-              },
-            },
-          },
-        },
-      });
-    }
-  }
-
   public async create(politician: any) {
-    await this.createIndex();
     await this.elasticsearchService.index({
       index: politiciansIndex,
       body: politician,
@@ -277,16 +156,5 @@ export class PoliticiansService {
       },
     });
     return null;
-  }
-
-  public async bulkCreate(politicians) {
-    await this.createIndex();
-    const { errors, took } = await this.elasticsearchService.bulk({
-      operations: politicians,
-    });
-    if (errors) {
-      return false;
-    }
-    return took;
   }
 }
